requirements.txt
aiogram==3.24.0
fastapi==0.115.12
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
greenlet==3.3.2
h11==0.16.0
idna==3.11
outcome==1.3.0.post0
psycopg2-binary==2.9.11
pycparser==3.0
PySocks==1.7.1
python-dotenv==1.2.1
requests==2.32.5
selenium==4.41.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.46
trio==0.33.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.35.0
websocket-client==1.9.0
wsproto==1.3.2
